+++ example/lib/drawer.dart
@@ -3,6 +3,7 @@ import 'package:flutter_map_marker_popup_example/popup_option_controls.dart';
 import 'package:flutter_map_marker_popup_example/simple_map_with_popups.dart';
 
 import 'popup_outside_of_map.dart';
+import 'selected_marker_builder.dart';
 
 Widget _buildMenuItem(
   BuildContext context,
@@ -44,6 +45,12 @@ Drawer buildDrawer(BuildContext context, String currentRoute) {
           context,
           const Text('Simple map with popups'),
           SimpleMapWithPopups.route,
+          currentRoute,
+        ),
+        _buildMenuItem(
+          context,
+          const Text('Selected marker builder'),
+          SelectedMarkerBuilder.route,
           currentRoute,
         ),
         _buildMenuItem(