+++ example/lib/main.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map_marker_popup_example/popup_option_controls.dart';
+import 'package:flutter_map_marker_popup_example/selected_marker_builder.dart';
 import 'package:flutter_map_marker_popup_example/simple_map_with_popups.dart';
 
 import 'popup_outside_of_map.dart';
@@ -18,6 +19,7 @@ class MyApp extends StatelessWidget {
       routes: <String, WidgetBuilder>{
         PopupOptionControls.route: (context) => const PopupOptionControls(),
         SimpleMapWithPopups.route: (context) => const SimpleMapWithPopups(),
+        SelectedMarkerBuilder.route: (context) => const SelectedMarkerBuilder(),
         PopupOutsideOfMap.route: (context) => const PopupOutsideOfMap(),
       },
     );