+++ lib/src/popup_marker_layer_options.dart
@@ -30,6 +30,11 @@ class PopupMarkerLayerOptions extends MarkerLayerOptions {
   /// animation.
   final PopupAnimation? popupAnimation;
 
+  /// An optional builder which, if provided, will be used to build markers
+  /// which are selected.
+  final Widget Function(BuildContext context, Marker marker)?
+      selectedMarkerBuilder;
+
   /// Setting a [MarkerCenterAnimation] will cause the map to be centered on
   /// a marker when it is tapped. Defaults to not centering on the marker.
   final MarkerCenterAnimation? markerCenterAnimation;
@@ -84,6 +89,7 @@ class PopupMarkerLayerOptions extends MarkerLayerOptions {
     this.initiallySelectedMarkers = const [],
     this.popupSnap = PopupSnap.markerTop,
     this.popupAnimation,
+    this.selectedMarkerBuilder,
     this.markerCenterAnimation,
     MarkerTapBehavior? markerTapBehavior,
     this.onPopupEvent,